+++ SIL.Windows.Forms.Archiving/Generic/ArchivingAccess.cs
@@ -1,0 +1,16 @@
+﻿
+namespace SIL.Windows.Forms.Archiving.Generic
+{
+	/// <summary></summary>
+	public class ArchivingAccess
+	{
+		/// <summary />
+		public string AccessProtocolName { get; set; }
+
+		/// <summary />
+		public string DateAvailable { get; set; }
+
+		/// <summary />
+		public string Owner { get; set; }
+	}
+}